+++ bin/make-bom.js
@@ -29,7 +29,7 @@ const xmlFormat = require('prettify-xml')
 
 const bomHelpers = require('../src')
 const program = require('../package.json')
-const Component = require('../src/model/Component')
+const Component = require('../src/models/Component')
 
 const EXIT_SUCCESS = 0
 const EXIT_INVALID = 1
