+++ src/models/AttachmentText.js
