+++ src/models/Component.js
