+++ src/models/CycloneDXObject.js
