+++ src/models/Hash.js
