+++ src/models/HashList.js
