+++ src/models/Swid.js
